requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests==2.28.1
python-dotenv==0.20.0
python-magic==0.4.24
airtable-python-wrapper==0.15.3
Flask-SocketIO
flask_sqlalchemy
